e2e/pyproject.toml
[project]
name = "clearmoney-e2e"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.3",
    "pytest-playwright>=0.7",
    "pytest-timeout>=2.3",
    "psycopg[binary]>=3.2",
]

[tool.pytest.ini_options]
addopts = "-x -vv"
base_url = "http://localhost:8765"
timeout = 30
# Serial execution — tests share DB state
# (equivalent to workers: 1 in playwright.config.ts)
